Refactor into folders.
