+++ esphome/components/neokey/binary_sensor/binary_sensor.py
@@ -2,7 +2,7 @@ import esphome.codegen as cg
 import esphome.config_validation as cv
 from esphome.components import binary_sensor
 from esphome.const import CONF_KEY
-from . import neokey_ns, NeoKeyComponent, CONF_NEOKEY_ID
+from .. import neokey_ns, NeoKeyComponent, CONF_NEOKEY_ID
 
 NeoKeyBinarySensor = neokey_ns.class_("NeoKeyBinarySensor", binary_sensor.BinarySensor)
